split inventory panel
