+++ crates/client/src/main.rs
@@ -52,7 +52,10 @@ use crate::{
     sprite_loader::{sprite_block_for_spriteref, sprite_block_for_tile, Scale, SpriteLoader},
     ui::{
         centered_rect,
-        panels::{render_credits_panel, render_help_panel, render_quit_panel},
+        panels::{
+            get_player_inventory, render_credits_panel, render_help_panel,
+            render_inventory_list_only, render_inventory_panel, render_quit_panel,
+        },
     },
 };
 use lithicrivers_core::model::body::{Body, BodyPart, BodyPartState};
@@ -155,43 +158,6 @@ struct App {
     boot_complete: bool,
     // Help panel state
     help_scroll: u16,
-}
-
-fn render_inventory_list_only(f: &mut Frame, app: &mut App, area: Rect) {
-    // Build list with selection highlight (same as right side of render_inventory_panel)
-    let mut list_lines: Vec<Line<'static>> = Vec::new();
-    if let Some(e) = app.game.res.player_entity {
-        if let Ok(inv) = app.game.world.get::<&InvComp>(e) {
-            if inv.slots.is_empty() {
-                list_lines.push(Line::from(Span::raw("(Empty)")));
-                app.inv_selected = 0;
-            } else {
-                if app.inv_selected >= inv.slots.len() {
-                    app.inv_selected = inv.slots.len() - 1;
-                }
-                for (i, s) in inv.slots.iter().enumerate() {
-                    let label = format!("{} x{}", itemkind_name(s.kind), s.qty);
-                    if i == app.inv_selected {
-                        list_lines.push(Line::from(Span::styled(
-                            label,
-                            Style::default().fg(Color::Yellow),
-                        )));
-                    } else {
-                        list_lines.push(Line::from(Span::raw(label)));
-                    }
-                }
-            }
-        } else {
-            list_lines.push(Line::from(Span::raw("(No Inventory component)")));
-        }
-    } else {
-        list_lines.push(Line::from(Span::raw("(No player)")));
-    }
-
-    let list_para = Paragraph::new(list_lines)
-        .alignment(Alignment::Left)
-        .block(Block::default().borders(Borders::ALL).title("Inventory"));
-    f.render_widget(list_para, area);
 }
 
 #[derive(Debug, Clone)]
@@ -1559,92 +1525,6 @@ fn render_game_view(f: &mut Frame, app: &mut App, area: Rect) {
     f.render_widget(paragraph, area);
 }
 
-fn render_inventory_panel(f: &mut Frame, app: &mut App, area: Rect) {
-    // Split: left art/details, right list
-    let chunks = Layout::default()
-        .direction(Direction::Horizontal)
-        .constraints([Constraint::Length(28), Constraint::Min(20)])
-        .split(area);
-
-    // Build list with selection highlight
-    let mut list_lines: Vec<Line<'static>> = Vec::new();
-    let mut selected_kind: Option<ItemKind> = None;
-    let mut selected_qty: u32 = 0;
-    if let Some(e) = app.game.res.player_entity {
-        if let Ok(inv) = app.game.world.get::<&InvComp>(e) {
-            if inv.slots.is_empty() {
-                list_lines.push(Line::from(Span::raw("(Empty)")));
-                app.inv_selected = 0;
-            } else {
-                if app.inv_selected >= inv.slots.len() {
-                    app.inv_selected = inv.slots.len() - 1;
-                }
-                for (i, s) in inv.slots.iter().enumerate() {
-                    let label = format!("{} x{}", itemkind_name(s.kind), s.qty);
-                    if i == app.inv_selected {
-                        selected_kind = Some(s.kind);
-                        selected_qty = s.qty;
-                        list_lines.push(Line::from(Span::styled(
-                            label,
-                            Style::default().fg(Color::Yellow),
-                        )));
-                    } else {
-                        list_lines.push(Line::from(Span::raw(label)));
-                    }
-                }
-            }
-        } else {
-            list_lines.push(Line::from(Span::raw("(No Inventory component)")));
-        }
-    } else {
-        list_lines.push(Line::from(Span::raw("(No player)")));
-    }
-
-    // Left: art + description if any selection
-    let left_block = Block::default().borders(Borders::ALL).title("Item");
-    let left_inner = left_block.inner(chunks[0]);
-    let mut left_lines: Vec<Line<'static>> = Vec::new();
-    if let Some(kind) = selected_kind {
-        let (category, sprite_name) = ("items", itemkind_sprite_name(kind));
-        let sd = app.sprite_loader.load_sprite(sprite_name, category);
-        // Art
-        if let Some(block) = sd.art12x8_sprites.first() {
-            let color = parse_hex_color(&sd.color);
-            for row in block.split('\n') {
-                let sp = match color {
-                    Some(c) => Span::styled(row.to_string(), Style::default().fg(c)),
-                    None => Span::raw(row.to_string()),
-                };
-                left_lines.push(Line::from(sp));
-            }
-        }
-        // Spacer and description
-        left_lines.push(Line::from(""));
-        left_lines.push(Line::from(Span::styled(
-            format!("{} (x{})", itemkind_name(kind), selected_qty),
-            Style::default().fg(Color::Cyan),
-        )));
-        left_lines.push(Line::from(""));
-        left_lines.push(Line::from(Span::raw(sd.description.clone())));
-        left_lines.push(Line::from(""));
-        left_lines.push(Line::from(Span::raw("Keys:")));
-        left_lines.push(Line::from(Span::raw("  d=drop")));
-        left_lines.push(Line::from(Span::raw("  .=duplicate")));
-        left_lines.push(Line::from(Span::raw("  x=destroy")));
-    } else {
-        left_lines.push(Line::from(Span::raw("Select an item")));
-    }
-    let left_para = Paragraph::new(left_lines).alignment(Alignment::Left);
-    f.render_widget(left_para, left_inner);
-    f.render_widget(left_block, chunks[0]);
-
-    // Right: list panel
-    let list_para = Paragraph::new(list_lines)
-        .alignment(Alignment::Left)
-        .block(Block::default().borders(Borders::ALL).title("Inventory"));
-    f.render_widget(list_para, chunks[1]);
-}
-
 fn render_message_log(f: &mut Frame, app: &mut App, area: Rect) {
     let mut lines: Vec<Line> = Vec::new();
     // Account for the border (top+bottom) since Paragraph has a Block
@@ -1813,23 +1693,6 @@ fn render_body_panel(f: &mut Frame, app: &mut App, area: Rect) {
 }
 
 // Helper function to get player's inventory as a HashMap
-fn get_player_inventory(app: &App) -> HashMap<ItemKind, u32> {
-    let mut inventory = HashMap::new();
-
-    if let Some(player_e) = app.game.res.player_entity {
-        if let Ok(inv) = app
-            .game
-            .world
-            .get::<&lithicrivers_core::components::Inventory>(player_e)
-        {
-            for item in &inv.slots {
-                *inventory.entry(item.kind).or_insert(0) += item.qty;
-            }
-        }
-    }
-
-    inventory
-}
 
 fn render_crafting_panel(f: &mut Frame, app: &mut App, area: Rect) {
     let block = Block::default().borders(Borders::ALL).title("Crafting");
