+++ crates/client/src/ui/panels/mod.rs
@@ -1,7 +1,9 @@
 pub mod credits;
 pub mod help;
+pub mod inventory;
 pub mod quit;
 
 pub use credits::render_credits_panel;
 pub use help::render_help_panel;
+pub use inventory::{get_player_inventory, render_inventory_list_only, render_inventory_panel};
 pub use quit::render_quit_panel;
